keep notifyPreferenceSchema inside users
