+++ backend/lib/endpoints/schema/users.js
@@ -67,6 +67,44 @@ const getUserByIdSchema = {
   params: strictSchema().prop("userId", S.string().required()),
 };
 
+const notifyPreferenceSchema = strictSchema()
+  .prop(
+    "message",
+    S.object()
+      .prop("instant", S.boolean().default(true))
+      .prop("daily", S.boolean().default(false))
+      .prop("weekly", S.boolean().default(false))
+      .prop("biweekly", S.boolean().default(false)),
+  )
+  .prop(
+    "like",
+    S.object()
+      .prop("instant", S.boolean().default(false))
+      .prop("daily", S.boolean().default(false))
+      .prop("weekly", S.boolean().default(true))
+      .prop("biweekly", S.boolean().default(false)),
+  )
+  .prop(
+    "comment",
+    S.object()
+      .prop("instant", S.boolean().default(false))
+      .prop("daily", S.boolean().default(true))
+      .prop("weekly", S.boolean().default(false))
+      .prop("biweekly", S.boolean().default(false)),
+  )
+  .prop(
+    "post",
+    S.object()
+      .prop("instant", S.boolean().default(false))
+      .prop("daily", S.boolean().default(true))
+      .prop("weekly", S.boolean().default(false))
+      .prop("biweekly", S.boolean().default(false)),
+  );
+
+const updateNotifyPrefsSchema = {
+  body: strictSchema().prop("notifyPrefs", notifyPreferenceSchema),
+};
+
 const updateUserSchema = {
   body: strictSchema()
     .prop("about", S.string().maxLength(160))
@@ -77,7 +115,8 @@ const updateUserSchema = {
     .prop("needs", needsSchema)
     .prop("objectives", objectivesSchema)
     .prop("photo", S.string().pattern(URL_REGEX))
-    .prop("urls", urlsSchema),
+    .prop("urls", urlsSchema)
+    .prop("notifyPrefs", notifyPreferenceSchema),
 };
 
 module.exports = {
@@ -85,4 +124,5 @@ module.exports = {
   createUserSchema,
   getUserByIdSchema,
   updateUserSchema,
+  updateNotifyPrefsSchema,
 };
